+++ EasyPost.Extensions/Testing/DummyData/Insurance.cs
@@ -2,13 +2,16 @@ using EasyPost.Extensions.ServiceMethodExtensions;
 
 namespace EasyPost.Extensions.Testing.DummyData;
 
+/// <summary>
+///     Functions for dummy <see cref="EasyPost.Models.API.Insurance"/>.
+/// </summary>
 public abstract class Insurance : DummyDataCreator
 {
     /// <summary>
     ///     Create a dummy <see cref="EasyPost.Models.API.Insurance"/>.
     /// </summary>
     /// <param name="client">The <see cref="EasyPost.Client"/> to make the API call with.</param>
-    /// <returns>A <see cref="Insurance"/> object.</returns>
+    /// <returns>A <see cref="EasyPost.Models.API.Insurance"/> object.</returns>
     public static async Task<EasyPost.Models.API.Insurance> CreateInsurance(Client client)
     {
         var parameters = new Parameters.V2.Insurance.Create