+++ EasyPost.Extensions/Testing/DummyData/Trackers.cs
@@ -2,10 +2,11 @@ using EasyPost.Models.API;
 
 namespace EasyPost.Extensions.Testing.DummyData;
 
+/// <summary>
+///     Functions for dummy <see cref="Tracker"/>s.
+/// </summary>
 public abstract class Trackers : DummyDataCreator
 {
-    private static string JsonFile => "assets/dummy_data/trackers.json";
-
     /// <summary>
     ///     Create a dummy <see cref="Tracker"/>.
     /// </summary>
@@ -13,13 +14,9 @@ public abstract class Trackers : DummyDataCreator
     /// <returns>A <see cref="Tracker"/> object.</returns>
     public static async Task<Tracker> CreateTracker(Client client)
     {
-        var data = GetRandomMapsFromJsonFile(JsonFile, 1, true);
-        var trackerData = data[0];
-        
-        // THIS IS SO INCONSISTENT, DICTIONARY VS EXPLICIT PARAMS
-        var code = trackerData["tracking_code"] as string;
-        var carrier = trackerData["carrier"] as string;
+        var carrier = Carriers.GetCarrier();
+        var trackingCode = $"EZ{Internal.Random.RandomStringOfLength(12)}";
 
-        return await client.Tracker.Create(carrier, code);
+        return await client.Tracker.Create(carrier, trackingCode);
     }
 }